添加权限详细参数 增加菜单栏 按钮 目录  添加强大的vue.js

+++ initial_version_of_springboot/src/main/java/com/fc/test/mapper/auto/TsysPremissionMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TsysPremissionMapper {
-    int countByExample(TsysPremissionExample example);
+    long countByExample(TsysPremissionExample example);
 
     int deleteByExample(TsysPremissionExample example);
